requirements.txt
pyramid
pyramid-jinja2
pyramid-debugtoolbar
faker
uwsgi
